% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Achehunting}
\alias{Achehunting}
\title{Human forager hunting returns data}
\format{\enumerate{
  \item month : Month of record
  \item day : Day of record
  \item year : Year of record
  \item id : Identifier of individual man
  \item age : Man's age at time of record
  \item kg.meat : Kilograms of meat returned from hunt
  \item hours : Duration in hours of hunting trip
  \item datatype : 1
if duration of trip known, 3 otherwise }}
\source{
Hill and Kintigh. 2009. Current Anthropology 50:369-377.
}
\usage{
Achehunting
}
\description{
Hunting returns of individual Ache men, 1981 to 2007.
}
