% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bangladesh}
\alias{bangladesh}
\title{Bangladesh contraceptive use data}
\format{\enumerate{
  \item woman : ID number for each woman in sample
  \item district : Number for each district
  \item use.contraception : 0/1 indicator of contraceptive use
  \item living.children : Number of living children
  \item age.centered : Centered age
  \item urban : 0/1 indicator of urban context }}
\source{
Bangladesh Fertility Survey, 1989
}
\usage{
bangladesh
}
\description{
Contraceptive use data from 1934 Bangladeshi women.
}
