% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{chainmode}
\alias{chainmode}
\title{Find mode of a continuous density estimate}
\usage{
chainmode(chain, ...)
}
\arguments{
\item{chain}{Values, e.g. sampled from a posterior via MCMC}

\item{...}{Optional arguments passed to density calculation}
}
\description{
Returns estimated mode of a density computed from samples.
}
\details{
This function just finds the x value that maximizes the y density in the
density estimate.
}
\author{
Richard McElreath
}
