% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{chimpanzees}
\alias{chimpanzees}
\title{Chimpanzee prosocialty experiment data}
\format{\enumerate{
  \item actor : name of actor
  \item recipient : name of recipient (NA for partner absent condition)
  \item condition : partner absent (0), partner present (1)
  \item block : block of trials (each actor x each recipient 1 time)
  \item trial : trial number (by chimp = ordinal sequence of
    trials for each chimp, ranges from 1-72; partner present trials were
    interspersed with partner absent trials)
  \item prosocial_left : 1 if prosocial (1/1) option was on left
  \item chose_prosoc : choice chimp made (0 = 1/0 option, 1 = 1/1 option)
  \item pulled_left : which side did chimp pull (1 = left, 0 = right) }}
\source{
Silk et al. 2005. Nature 437:1357-1359.
}
\usage{
chimpanzees
}
\description{
Data from behavior trials in a captive group of chimpanzees, housed in
Lousiana. From Silk et al. 2005. Nature 437:1357-1359.
}
\author{
Richard McElreath
}
