% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeftab.r
\docType{class}
\name{coeftab-class}
\alias{coeftab-class}
\title{An S4 class for storing a coefficient table}
\description{
An S4 class for storing a coefficient table
}
\section{Slots}{

\describe{
\item{\code{coefs}}{A matrix of coefficients}

\item{\code{se}}{A matrix of standard errors}

\item{\code{nobs}}{Number of observations}

\item{\code{AIC}}{AIC}

\item{\code{digits}}{A numeric indicating how many digits to display}

\item{\code{width}}{a numberic indicating width for formatting output}
}}

