% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeftab.r
\name{coeftab}
\alias{coeftab}
\title{Coefficient tables}
\usage{
coeftab(..., se = FALSE, se.inside = FALSE, nobs = TRUE, digits = 2,
  width = 7, rotate = FALSE)
}
\arguments{
\item{...}{A series of fit models, separated by commas}

\item{se}{Include standard errors in table?}

\item{se.inside}{Print standard errors in same cell as estimates}

\item{nobs}{Print number of observations for each model?}

\item{digits}{Number of digits to round numbers to}

\item{rotate}{If TRUE, rows are models and columns are coefficients}
}
\description{
Returns a table of model coefficients in rows and models in columns.
}
\details{
This function provides a way to compare estimates across models.
}
\author{
Richard McElreath
}
