% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareIC-class.R
\docType{class}
\name{compareIC-class}
\alias{compareIC-class}
\alias{compareIC}
\title{The class compareIC}
\description{
Output of the function \code{\link{compare}}. It contains
the class \code{data.frame}.
}
\section{Slots}{

\describe{
\item{\code{dSE}}{matrix of standard errors of differences in IC between pairs of models}

\item{\code{.Data}}{contains the actual data frame with the results}
}}

