% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Crofoot}
\alias{Crofoot}
\title{Capuchin monkey contests}
\format{\enumerate{
  \item focal : ID of focal group
  \item other : ID of other group
  \item dyad : ID of specific dyad pair of groups
  \item win : 1 if focal won contest, 0 if other won
  \item dist_focal : Distance in meters of focal group from
    the center of its home range
  \item dist_other : Distance in
meters of other group from the center of its home range
  \item n_focal : Number of individuals in focal group
  \item n_other : Number of individuals in other group
  \item m_focal : Number of males in focal group
  \item m_other : Number of males in other group
  \item f_focal : Number of females in focal group
  \item f_other : Number of females in other group }}
\source{
M.C. Crofoot, I.C. Gilby, M.C. Wikelski, and R.W. Kays. 2008.
PNAS 105:577--581.
}
\usage{
Crofoot
}
\description{
Data on features and outcomes of intergroup contests among territorial
groups of capuchin monkey (Cebus capucinus). Each row contains a single
contest between two groups.
}
