% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densify.R
\name{densify}
\alias{densify}
\title{Turn "rasterized" kernel into "rasterized" density function}
\usage{
densify(x, y)
}
\arguments{
\item{x, y}{numerical vectors of equal length giving the (x, y) coordinates
of a kernel function \code{k} with \code{k(x) = y}.}
}
\value{
a rescaled version of \code{y} that turns the kernel into a pdf.
}
\description{
Turn "rasterized" kernel into "rasterized" density function
}
\examples{
x <- seq(0, 1, by = 0.05)
y1 <- (x - 0.5)^2
y2 <- densify(x, y1)
if (require(lattice)) {
  xyplot(y1 + y2 ~ x, type = "l")
}
}
