% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Dinasaurs}
\alias{Dinasaurs}
\title{Dinosaur growth data}
\format{\enumerate{
  \item age : Estimated age of specimen at death, in years
  \item mass : Estimated body mass at death, in kilograms
  \item species : Species name
  \item sp_id : Identification number of species }}
\source{
Erickson, Rogers, Yerby. 2001. Dinosaurian growth patterns and
rapid avian growth rates. Nature 412:429-433.
}
\usage{
Dinasaurs
}
\description{
Mass by age data for 6 species of dinosaur.
}
