\name{Dinosaurs}
\alias{Dinosaurs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Dinosaur growth data}
\description{
  Mass by age data for 6 species of dinosaur.
}
\usage{
data(Dinosaurs)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
    \enumerate{
        \item age : Estimated age of specimen at death, in years
        \item mass : Estimated body mass at death, in kilograms
        \item species : Species name
        \item sp_id : Identification number of species
    }
}
\value{
}
\references{Erickson, Rogers, Yerby. 2001. Dinosaurian growth patterns and rapid avian growth rates. Nature 412:429-433.}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

