% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Dissertations}
\alias{Dissertations}
\title{Dissertation page length data}
\format{\enumerate{
  \item pages: number of pages
  \item major: department/program of dissertation
  \item year: year of filing }}
\source{
Data originally parsed by...
}
\usage{
Dissertations
}
\description{
Page lengths of 3037 dissertations filed between 2006 and 2014 at the
University of Minnesota.
}
