% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.r
\name{dordlogit}
\alias{dordlogit}
\alias{rordlogit}
\alias{pordlogit}
\title{Ordered categorical log-odds probability density}
\usage{
dordlogit(x, phi, a, log = FALSE)
}
\arguments{
\item{x}{Integer values to compute densities or probabilies of}

\item{phi}{Linear model of log-odds}

\item{a}{Vector of log-odds intercepts}

\item{log}{If \code{TRUE}, returns log-probability instead of probability}
}
\description{
Functions for computing density, cumulative probability and producing random
samples from an ordered categorical probability density.
}
\details{
These functions provide for common density calculations for the ordered
categorical probability density commonly known as an "ordered logit" or
"ordered logistic." This is the same probability density assumed by the
\code{polr} function (library \code{MASS}).
}
\author{
Richard McElreath
}
