% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.r
\name{dzagamma2}
\alias{dzagamma2}
\title{Zero-augmented gamma probability density}
\usage{
dzagamma2(x, prob, mu, scale, log = FALSE)
}
\arguments{
\item{x}{Values to compute densities for}

\item{prob}{Probability of a zero}

\item{mu}{Mean of gamma distribution}

\item{scale}{Scale parameter (same as \code{scale} in \code{\link{dgamma}})}

\item{log}{If \code{TRUE}, returns log-density instead of density}
}
\description{
Function for computing density from a zero-augmented gamma probability
distribution.
}
\details{
This distribution is defined as a finite mixture of zeros and strictly
positive gamma distributed values, where \code{prob} determines the weight
of the zeros. As such, the probability of a zero is \code{prob}, and the
probability of a non-zero value \code{x} is \code{(1-prob)*dgamma( x ,
mu/scale , scale )}.
}
\examples{

}
\author{
Richard McElreath
}
