% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Fish}
\alias{Fish}
\title{Fishing data}
\format{\enumerate{
  \item fish_caught : Number of fish caught during visit
  \item livebait : Whether or not group used livebait to fish
  \item camper : Whether or not group had a camper
  \item persons : Number of adults in group
  \item child : Number of children in group
  \item hours : Number of hours group spent in park }}
\usage{
Fish
}
\description{
Fishing data from visitors to a national park.
}
