% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{foxes}
\alias{foxes}
\title{Urban foxes}
\format{\enumerate{
  \item group : ID of group
  \item avgfood : Average available food in group's territory
  \item groupsize : Size of each group
  \item area : Area of group territory
  \item weight : Body weight of individual fox }}
\source{
Modified from an example in Grafen and Hails.
}
\usage{
data(foxes)
}
\description{
Data on urban fox groups in England.
}
