% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precis.r
\name{ggplot.precis}
\alias{ggplot.precis}
\title{ggplot2 plot of precis of models fit with rethinking package}
\usage{
ggplot.precis(x, params, col_ci = "black", lab = "value", flip = TRUE,
  dot_size = 1, line_size = 0.5, ...)
}
\arguments{
\item{x}{a precis object}

\item{params}{character vectro of parameters to include}

\item{col_ci}{color of interval}

\item{flip}{logical indicating whther x- and y- axes should be swapped.
If TRUE (the default), then parameters will be listed along the y-axis.}

\item{dot_size}{size of dots}

\item{line_size}{size of lines}

\item{...}{additional arguments (currently ignored)}
}
\description{
ggplot2 plot of precis of models fit with rethinking package
}
