\name{Howell}
\alias{Howell1}\alias{Howell2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Howell !Kung demography data}
\description{
  Demographic data from Kalahari !Kung San people collected by Nancy Howell
}
\usage{
data(Howell1)
data(Howell2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
    \enumerate{
        \item height: Height in cm
        \item weight: Weight in kg
        \item age: Age in years
        \item male: Gender indicator
        \item age.at.death: If deceased, age at death
        \item alive: Indicator if still alive
    }
}
\value{
}
\references{Downloaded from https://tspace.library.utoronto.ca/handle/1807/10395}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

