% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Hurricanes}
\alias{Hurricanes}
\title{Hurricane fatalities and gender of names}
\format{\enumerate{
  \item name : Given name of hurricane
  \item year : Year of hurricane
  \item deaths : number of deaths
  \item category : Severity code for storm
  \item min_pressure : Minimum pressure, a measure of storm strength;
     low is stronger
  \item damage_norm : Normalized estimate of damage in dollars
  \item female : Indicator variable for female name
  \item femininity : 1-11 scale from totally masculine (1)
     to totally feminine (11) for name. Average of 9 scores from 9 raters.  }}
\source{
Jung et al. 2014. Female hurricanes are deadlier than male
hurricanes. PNAS.
}
\usage{
data(Hurricanes)
}
\description{
Data used in Jung et al 2014 analysis of effect of gender of name on
hurricane fatalities. Note that hurricanes Katrina (2005) and Audrey (1957)
were removed from the data.
}
