% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{image_xyz}
\alias{image_xyz}
\title{Heat map from equal length x,y,z vectors}
\usage{
image_xyz(x, y, z, ...)
}
\arguments{
\item{x}{vector of x values}

\item{y}{vector of y values}

\item{z}{vector of z values}

\item{...}{other parameters to pass to \code{image}}
}
\description{
Provides an interface to use \code{image} by providing three equal length
vectors for x, y and z coordinates.
}
\details{
This function merely constructs a matrix suitable for \code{image}, using x,
y and z coordinates.
}
\seealso{
\code{\link{image}}
}
\author{
Richard McElreath
}
