% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Kline}
\alias{Kline}
\alias{Kline2}
\alias{islandsDistMatrix}
\alias{islands}
\title{Oceanic islands data}
\format{\enumerate{
  \item culture : Name of island culture
  \item population : Historical population size
  \item contact : low or high contact rate with other islands
  \item total_tools : number of tools in historical tool kit
  \item mean_TU : another measure of tool complexity
  \item lat : latitude of island
  \item lon : longitude of island
  \item lon2 : longitude coded for ease of plotting
  \item logpop : natural logarithm of population }}
\source{
Kline, M.A. and R. Boyd. 2010. Proc R Soc B 277:2559--2564.
}
\usage{
data(Kline)
}
\description{
Data on historic tool complexity and demography in various Oceanic islands
societies. There are three data tables: (1) \code{Kline} is the basic data
table; (2) \code{Kline2} contains latitude and longitude columns; and (3)
\code{islandsDistMatrix} is a matrix of pairwise distances between islands,
in thousands of kilometers.
}
