% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{KosterLeckie}
\alias{KosterLeckie}
\title{Gift exchange data}
\format{\code{kl_dyads}: \enumerate{

  \item hidA : household ID for A in dyad
  \item hidB : household ID for B in dyad
  \item did : dyad ID number
  \item giftsAB : count of gifts from A to B
  \item giftsBA : count of gifts from B to A
  \item offset : relative rate of contact in dyad
  \item drel1 :
  \item drel2 :
  \item drel3 :
  \item drel4 :
  \item dlndist :
  \item dass :
  \item d0125 :
  }
  \code{kl_households}: \enumerate{
  \item hid : household ID
  \item hgame :
  \item hfish :
  \item hpigs :
  \item hwealth :
  \item hpastor : }}
\source{
Koster and Leckie (2014) Food sharing networks in lowland
Nicaragua: An application of the social relations model to count data.
Social Networks.
}
\usage{
KosterLeckie
}
\description{
Data on household gift exchanges from Koster and Leckie. There are two data
frames: \code{kl_dyads} and \code{kl_households}.
}
