\name{Laffer}
\alias{Laffer}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Corporate tax rates and tax revenue for 20 nations}
\description{
  Data on 2004 corporate tax rates and tax revenue for 29 nations, as published in The Wall Street Journal in 2007.
}
\usage{
data(Laffer)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
    \enumerate{
        \item tax_rate: Corporate tax rate
        \item tax_revenue: Tax revenue
    }
}
\value{
}
\references{"We're Number One, Alas" 2007. The Wall Street Journal.}
\author{Richard McElreath}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

