% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.r
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Rethinking utility functions}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{vector of values}
}
\description{
Numerically stable computation of log sums of exponentiated values.
}
\details{
This function is used by \code{WAIC} to compute the log average probability
used in the formula for WAIC. In that context, a vector of log-probabilities
is passed to \code{log_sum_exp}, obviating the need to explicitly
exponentiate. This helps to avoid rounding errors that occur when working
with direct probabilities.
}
\examples{


}
\author{
Richard McElreath
}
