% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-class.r
\docType{class}
\name{map-class}
\alias{map-class}
\title{An S4 class to represent a MAP fit by quadratic approximation using \code{\link{map}()}}
\description{
An S4 class to represent a MAP fit by quadratic approximation using \code{\link{map}()}
}
\section{Slots}{

\describe{
\item{\code{call}}{The original call to \code{\link{map}()}}

\item{\code{coef}}{The estimated posterior modes}

\item{\code{vcov}}{Variance-covariance matrix}

\item{\code{optim}}{List returned by \code{\link{optim}()}}

\item{\code{data}}{The data}

\item{\code{formula}}{Formula list}

\item{\code{formula_parsed}}{Parsed version of \code{formula} as a list
of character.}

\item{\code{fminuslogl}}{Negative log-likelihood function that accepts a vector of
parameter values}

\item{\code{links}}{List of links}

\item{\code{start}}{List of starting values for search algorithm}
}}

