% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{map2stan-class}
\alias{map2stan-class}
\alias{show,map2stan-method}
\alias{pairs,map2stan-method}
\alias{plot,map2stan-method}
\alias{extract.samples,map2stan-method}
\alias{stancode,map2stan-method}
\alias{WAIC,map2stan-method}
\alias{DIC,map2stan-method}
\alias{deviance,map2stan-method}
\alias{sim,map2stan-method}
\alias{logLik,map2stan-method}
\alias{nobs,map2stan-method}
\alias{summary,map2stan-method}
\alias{vcov,map2stan-method}
\title{Class \code{map2stan}: fitted map2stan Stan model}
\description{
This object contains an object of class \code{stanfit}, return by
\code{\link{stan}}, along with other information produced by a
\code{\link{map2stan}} model fit.
}
\section{Slots}{
 \describe{ \item{list("call")}{Original function call that
produced the fit}\item{:}{Original function call that produced the fit}
\item{list("model")}{Text of the Stan model used in sampling}\item{:}{Text
of the Stan model used in sampling} \item{list("stanfit")}{Object of class
\code{stanfit}}\item{:}{Object of class \code{stanfit}}
\item{list("coef")}{Vector of posterior means}\item{:}{Vector of posterior
means} \item{list("vcov")}{Diagonal matrix with parameter
variances}\item{:}{Diagonal matrix with parameter variances}
\item{list("data")}{Data used in sampling}\item{:}{Data used in sampling}
\item{list("start")}{Initial values used in sampling}\item{:}{Initial values
used in sampling} \item{list("pars")}{Vector of parameters returned by
sampling}\item{:}{Vector of parameters returned by sampling}
\item{list("formula")}{Original \code{alist} of formulas used to define
model}\item{:}{Original \code{alist} of formulas used to define model}
\item{list("formula_parsed")}{Parsed formula list. Used during compilation
and useful for debugging and class methods.}\item{:}{Parsed formula list.
Used during compilation and useful for debugging and class methods.} }
}

\seealso{
\code{\link{map2stan}}
}
\keyword{classes}
