% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{mcreplicate}
\alias{mcreplicate}
\title{Multi-core version of replicate}
\usage{
mcreplicate(n, expr, refresh = 0.1, mc.cores = 2)
}
\arguments{
\item{n}{Number of replications}

\item{expr}{Code to replicate}

\item{refresh}{Status update refresh interval}

\item{mc.cores}{Number of cores to use}
}
\description{
Uses the \code{parallel} library to distribute \code{\link{replicate}}
processing across cores.
}
\details{
This function uses \code{\link{mclapply}} to distribute replications across
cores. It then simplifies the result to an array.
}
\seealso{
\code{\link{mclapply}}, \code{\link{replicate}}
}
\author{
Richard McElreath
}
