\name{milk}
\alias{milk}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Primate milk data}
\description{
  Comparative primate milk composition data, from Table 2 of Hinde and Milligan. 2011. Evolutionary Anthropology 20:9-23.
}
\usage{
data(milk)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
  Returns a data frame containing 29 rows and 8 columns.
    \enumerate{
      \item clade: Broad taxonomic group
      \item species: Species name
      \item kcal.per.g: Kilocalories per gram of milk
      \item perc.fat: Percent fat
      \item perc.protein: Percent protein
      \item perc.lactose: Percent lactose
      \item mass: Body mass of mother, in kilograms
      \item neocortex.perc: Percent of brain mass that is neocortex
    }
}
\value{
}
\references{Hinde and Milligan. 2011. Evolutionary Anthropology 20:9-23.}
\author{Richard McElreath}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

