\name{Moralizing_gods}
\alias{Moralizing_gods}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Religious ideology and population growth}
\description{
  Historical data on human societies, their population sizes, and the present or absence of moralizing gods. Moralizing gods are defined as gods that are believed to enforce prosocial behavior through punishment or reward.
}
\usage{
data(Moralizing_gods)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
    \enumerate{
        \item polity: Name of society
        \item year: Calendar year of record
        \item population: Log population of polity
        \item moralizing_gods: Indicator of presence (1), absence (0), or unknown (NA)
        \item writing: Presence of written records
    }
}
\value{
}
\references{Whitehouse et al. 2019. Complex societies precede moralizing gods throughout world history. doi:10.1038/s41586-019-1043-4}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

