% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_norm_unstand.R
\name{normalize}
\alias{normalize}
\alias{standardize}
\alias{unstandardize}
\title{Standardize/normalize vectors}
\usage{
normalize(x)

standardize(x)

unstandardize(
  x,
  center = attr(x, "scaled:center"),
  scale = attr(x, "scaled:scale")
)
}
\arguments{
\item{x}{a numeric vector.}

\item{center}{the center of the original distribution. If not provided, it is read from the attribute \code{scaled:center}.}

\item{scale}{the scale of the original distribution. If not provided, it is read from the attribute \code{scaled:scale}}
}
\value{
for \code{normalize}: the normalized vector

for \code{standardize} a standardized vector with attributes \code{scaled:center} and \code{scaled:scale}, containing the center and scale used for standardization.

for \code{unstandardize}: a vector with the original values.
}
\description{
Standardization is done by substracting the mean and dividing by the standard deviation. It does the same as \code{\link{scale}}, but works on vectors instead of matrices. Normalization is done by substracting the minimum and dividing by the maximum of the resulting vector. This bounds the vector between 0 and 1.
}
\examples{

x <- rnorm(10, 5, 2)
# Normalization
xn <- normalize(x)
range(xn)
# Standardization
xs <- standardize(x)
xorig <- unstandardize(xs)

mean(xs)
sd(xs)
identical(x, xorig)

}
