\name{Panda_nuts}
\alias{Panda_nuts}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Chimpanzees cracking Panda nuts}
\description{
  Data on 22 western chimpanzees attempting to crack Panda nuts.
}
\usage{
data(Panda_nuts)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
    \enumerate{
        \item chimpanzee: ID of individual
        \item age: Age in years at time of observation
        \item sex: individual's sex
        \item hammer: Type of hammer (one of four types)
        \item nuts_opened: Number of nuts opened in session
        \item seconds: Duration of session in seconds
        \item help: Received help from another chimpanzee (yes or no)
    }
}
\value{
}
\references{Boesch, Bombjakova, Meier, and Mundry. 2019. "Learning curves and teaching when
acquiring nut-cracking in humans and chimpanzees". doi:10.1038/s41598-018-38392-8}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

