% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeftab.r
\docType{methods}
\name{plot,coeftab,ANY-method}
\alias{plot,coeftab,ANY-method}
\title{Plots of coefficient tables}
\usage{
\S4method{plot}{coeftab,ANY}(x, y, ...)
}
\arguments{
\item{x}{Object produced by \code{coeftab}}

\item{y}{\code{NULL} and unused. Required for compatibility with base
\code{plot}}

\item{pars}{Optional vector of parameter names or indexes to display. If
missing, all parameters shown.}

\item{col.ci}{Color to draw confidence intervals}

\item{by.model}{Cluster estimates by model instead of by parameter (default)}

\item{prob}{Probability mass for confidence intervals. Default is 0.95.}
}
\description{
Plots coefficient tables produced by \code{coeftab}, clustered either by
models or by parameter names.
}
\details{
This function plots the tabular output of \code{\link{coeftab}}, using a
\code{\link{dotchart}}. By default, estimates are grouped by parameter, with
a row for each model. Model's without a parameter still appear as a row, but
with no estimate. By setting \code{by.model=TRUE}, the dotchart will instead
be grouped by model, with each row being a parameter.

MAP estimates are displayed with percentile confidence (credible) intervals.
Default is 95\% intervals. Use \code{prob} to change the interval mass.
}
\seealso{
\code{\link{coeftab}}, \code{\link{dotchart}}
}
\author{
Richard McElreath
}
