% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{progbar}
\alias{progbar}
\title{Progress display}
\usage{
progbar(current, min = 0, max = 100, starttime, update.interval = 100,
  show.rate = FALSE)
}
\arguments{
\item{current}{Current loop index of process}

\item{min}{Minimum loop index, usually 0 or 1}

\item{max}{Maximum loop index}

\item{starttime}{Time stamp when process began, from \code{Sys.time}}

\item{update.interval}{How often to display progress}
}
\description{
Provides a progress display with estimated time remaining, assuming rate
constant process.
}
\details{
This function provides useful progress information and estimated time until
completion for long looped operations, like sampling from MCMC.
}
\author{
Richard McElreath
}
