% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_import-export.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{sampling}
\alias{reexports}
\alias{cpp_object_initializer}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Rcpp}{\code{\link[Rcpp]{cpp_object_initializer}}}

  \item{rstan}{\code{\link[rstan]{sampling}}}
}}

