\name{rethinking-internal}
\alias{rethinking-internal}

\alias{Rho}
\alias{arrowpline}
\alias{blank}
\alias{coeftab_show}
\alias{concat}
\alias{confint_quad}
\alias{covmat}
\alias{fclip}
\alias{flist_untag}
\alias{format_show}
\alias{htmlhelp}
\alias{inv_logit}
\alias{logistic}
\alias{make.grid}
\alias{map2stan.templates}
\alias{mcmcpairs}
\alias{multilogistic}
\alias{plot2y}
\alias{plotchains}
\alias{plotpost}
\alias{plp}
\alias{postlistprecis}
\alias{precis.plot}
\alias{precis.show}
\alias{precis.whitelist}
\alias{predict_ordlogit}
\alias{rangi2}
\alias{replicate2}
\alias{rrformat}
\alias{sd2}
\alias{se}
\alias{segmentsby}
\alias{set_nice_margins}
\alias{setcran}
\alias{show.naive.posterior}
\alias{softmax}
\alias{tracerplot}
\alias{var2}

\alias{AIC,ANY-method}
\alias{DIC,ANY-method}
\alias{WAIC,ANY-method}
\alias{coef,ANY-method}
\alias{deviance,ANY-method}
\alias{extract.samples,ANY-method}
\alias{link,ANY-method}
\alias{sim,ANY-method}
\alias{logLik,ANY-method}
\alias{nobs,ANY-method}
\alias{pairs,ANY-method}
\alias{plot,ANY-method}
\alias{show,ANY-method}
\alias{stancode,ANY-method}
\alias{summary,ANY-method}
\alias{vcov,ANY-method}

\title{Internal Functions and Methods}
\description{
These functions are not intended to be called directly. In the case of
methods, they documented elsewhere, either with the associated generic
function or with the class of the object that the method is defined for.
}
\keyword{internal}