% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rethinking-package.R
\docType{package}
\name{rethinking-package}
\alias{rethinking-package}
\alias{rethinking}
\title{Statistical Rethinking package}
\description{
This package accompanies a book and course on Bayesian data analysis,
featured MAP estimation through \code{\link{map}} and Hamiltonian Monte
Carlo through \code{\link{map2stan}}.
}
\details{
\tabular{ll}{ Package: \tab rethinking\cr Type: \tab Package\cr Version:
\tab 1.59\cr Date: \tab 24 Jun 2016\cr License: \tab GPL-3 \cr }
}
\references{
McElreath (2016). Statistical Rethinking: A Bayesian Course with
Examples in R and Stan. CRC Press.
}
\seealso{
\code{\link{map}}, \code{\link{map2stan}}
}
\author{
Richard McElreath
}
\keyword{package}
\keyword{rethinking}
