% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Rinder}
\alias{Rinder}
\title{Livestock mortality data}
\format{\enumerate{
  \item herd : identifier for individual herd
  \item stock : categorical small or large stock
  \item n : number of animals at beginning of observation period
  \item mortality : number of animals that died in observation period }}
\usage{
Rinder
}
\description{
Mortality counts for 11 herds of large and small livestock. Large livestock
are Zebu cattle, and small livestock are a mix of (fat-tailed, Maasai) sheep
and goats.
}
