% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.r
\docType{methods}
\name{sim,map-method}
\alias{sim,map-method}
\alias{sim}
\alias{sim-methods}
\alias{sim,map2stan-method}
\alias{sim,lm-method}
\alias{sim,map2stan-method}
\title{Simulate posterior observations}
\usage{
\S4method{sim}{map}(fit, data, n = 1000, post, ll = FALSE, refresh = 0.1,
  replace = list(), ...)

\S4method{sim}{map2stan}(fit, data, n = 1000, post, refresh = 0.1,
  replace = list(), ...)
}
\arguments{
\item{fit}{Object of class \code{map} or \code{map2stan}}

\item{data}{Optional list of data to compute predictions over}

\item{n}{Number of samples to use}

\item{post}{Samples from posterior. If missing, \code{sim} samples using
\code{n}}

\item{...}{Other parameters to pass to someone}
}
\description{
Simulates posterior observations for \code{map} and \code{map2stan} model
fits.
}
\details{
This function uses the model definition from a \code{map} or \code{map2stan}
fit to simulate outcomes that average over the posterior distribution. It
uses \code{\link{link}} internally to process any linear models. Then the
correspond random number function is used, as defined by the model's
likelihood.

The \code{rethinking} package defines a generic function \code{sim}, so
methods can be defined for other model fit classes. I might eventually build
methods for \code{lm} and \code{glm}.
}
\seealso{
\code{\link{link}},\code{\link{map}},\code{\link{map2stan}}
}
\author{
Richard McElreath
}
