% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeftab.r
\name{[.coeftab}
\alias{[.coeftab}
\title{Extract rows/columns from coeftab}
\usage{
\method{[}{coeftab}(x, ...)
}
\arguments{
\item{x}{a coeftab}

\item{...}{arguments passe to extractor for a matrix.}
}
\description{
Extract rows/columns from coeftab
}
