% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace-plots.R
\name{tracer_plot}
\alias{tracer_plot}
\alias{pairs.map2stan}
\alias{pairs.map}
\title{Trace plot of map2stan chains}
\arguments{
\item{x}{map or map2stan model fit}

\item{n}{Number of samples to show in scatter plots}

\item{alpha}{alpha transparency of scatter plots}

\item{cex}{Character expansion factor for scatter plots}

\item{pch}{Point type for scatter plots}

\item{adj}{\code{adj} argument for \code{\link{density}} estimate}

\item{pars}{Character vector of parameters to display}

\item{...}{Additional plot arguments}
}
\description{
Shows \code{pairs} plots for Stan samples produced by a \code{map} or
\code{map2stan} fit.
}
\details{
This is the default pairs plot method for both \code{\link{map}} and
\code{\link{map2stan}} model fits.
}
\author{
Richard McElreath
}
