% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2stan-class.r
\name{tracerplot}
\alias{tracerplot}
\title{Trace plot of map2stan chains}
\usage{
tracerplot(object, pars, col = rethink_palette, alpha = 1,
  bg = col.alpha("black", 0.15), ask = TRUE, window, n_cols = 3,
  max_rows = 5, ...)
}
\arguments{
\item{object}{map2stan model fit}

\item{pars}{Character vector of parameters to display}

\item{col}{Vector of colors to use for chains. Recycled.}

\item{alpha}{alpha transparency of chains}

\item{bg}{Background fill for warmup samples}

\item{ask}{Whether to pause for paging. Default is \code{TRUE}.}

\item{window}{Range of samples to display. Default is all samples.}

\item{n_cols}{Number of columns per page}

\item{max_rows}{Maximum number of rows per page}

\item{...}{Additional arguments to pass to plot commands}
}
\description{
Shows trace plots for Stan samples produced by a \code{map2stan} fit,
annotated by number of effective samples. Automatic paging and adjustable
number of columns and rows per page.
}
\details{
This is the default trace plot method for \code{\link{map2stan}} model fits.
}
\author{
Richard McElreath
}
