% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Trolly}
\alias{Trolly}
\title{Experimental data on ethical dilemmas}
\format{\enumerate{
  \item case: a code that combines treatment and story labels
  \item response: participant's rating of appropriateness of action in
     story
  \item order: integer order story appeared, within participant
  \item id: participant id (factor)
  \item age: participant's age in years
   \item male: participant's gender; 1 for male, 0 for female
  \item edu: participant's highest educational level
  \item action: treatment code for
story with action (1) or inaction (0)
  \item intention: treatment code for intent (1) or lack of intent (0)
  \item contact: treatmetn code for contact action (1) or lack of contact (0)
  \item story: factor label for basic scenario modified by treatments
  \item action2: alternative coding of action that is union of action and
     contact variables }}
\source{
Cushman et al. 2006. Psychological Science 17:1082--1089.
}
\usage{
data(Trolly)
}
\description{
These data comprise outcomes of experimental ethical dilemmas that are often
called 'trolley' problems. Data kindly provided by Fiery Cushman.
}
