% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{UFClefties}
\alias{UFClefties}
\title{Ultimate Fighting Championship matches}
\format{\enumerate{
  \item fight : Unique identifier for match
  \item episode : Identifier for UFC episode
  \item fight.in.episiode : Order of fight in episode
  \item fighter1.win : 1 if fighter 1 won the match; 0 if fight 2 won
  \item fighter1 : Unique identifier for fighter 1
  \item fighter2 : Unique identifier for fighter 2
  \item fighter1.lefty : 1 if fighter 1 was left handed; 0 otherwise
  \item fighter2.lefty : 1 if fighter 2 was left handed; 0 otherwise }}
\source{
Pollet et al. 2013. Animal Behaviour 86:839--843.
}
\usage{
UFClefties
}
\description{
Outcomes of televised Ultimate Fighting Championship (UFC) matches, as a
function of handedness of fighters. Data coded by Pollet et al (2013).
}
