% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Wines2012}
\alias{Wines2012}
\title{Wine tasting scores, Princeton NJ 2012}
\format{\enumerate{
  \item judge : Name of judge
  \item flight : white or red group of 10 wines
  \item wine : Wine ID
  \item score : Numerical score of wine
  \item wine.amer : 1 for American wines from New Jersey
  \item judge.amer : 1 for American judges }}
\source{
Raw data from http://www.liquidasset.com/report161.html
}
\usage{
data(Wines2012)
}
\description{
Numerical scores from 9 judges, both French and American, testing 20
different wines, both French and American. Judges were blind to the identity
of each wine during the testing.
}
