\name{xcoef}
\alias{se}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Rethinking utility functions}
\description{
  Various utility functions including in the book package.
}
\usage{
col.alpha( acol , alpha = 0.2 )
}
%- maybe also 'usage' for other objects documented here.
\references{McElreath 2011, Statistical Rethinking.}
\author{Richard McElreath}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

