\name{Achehunting}
\alias{Achehunting}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Human forager hunting returns data}
\description{
  Hunting returns of individual Ache men, 1981 to 2007.
}
\usage{
data(Achehunting)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
    \enumerate{
        \item month : Month of record
        \item day : Day of record
        \item year : Year of record
        \item id : Identifier of individual man
        \item age : Man's age at time of record
        \item kg.meat : Kilograms of meat returned from hunt
        \item hours : Duration in hours of hunting trip
        \item datatype : 1 if duration of trip known, 3 otherwise
    }
}
\value{
}
\references{Hill and Kintigh. 2009. Current Anthropology 50:369-377.}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

