\name{bangladesh}
\alias{bangladesh}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Bangladesh contraceptive use data}
\description{
  Contraceptive use data from 1934 Bangladeshi women.
}
\usage{
data(bangladesh)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
    \enumerate{
        \item woman : ID number for each woman in sample
        \item district : Number for each district
        \item use.contraception : 0/1 indicator of contraceptive use
        \item living.children : Number of living children
        \item age.centered : Centered age
        \item urban : 0/1 indicator of urban context
    }
}
\value{
}
\references{Bangladesh Fertility Survey, 1989}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

