% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeftab-class.R
\docType{class}
\name{coeftab-class}
\alias{coeftab-class}
\title{Class coeftab}
\description{
The class represents a table with coefficients.
}
\section{Slots}{

\describe{
\item{\code{coefs}}{a matrix with the coefficients}

\item{\code{se}}{a matrix with the standard errors}

\item{\code{nobs}}{a numeric vector with the number of observations}

\item{\code{AIC}}{the AIC.}

\item{\code{digits}}{the number of digits used to round numbers}

\item{\code{width}}{the minimum field width, see \code{\link{format}}}
}}

