% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeftab.r
\name{coeftab}
\alias{coeftab}
\title{Coefficient tables}
\usage{
coeftab(
  ...,
  se = FALSE,
  se.inside = FALSE,
  nobs = TRUE,
  digits = 2,
  width = 7,
  rotate = FALSE
)
}
\arguments{
\item{...}{a series of fit models, separated by commas}

\item{se}{if \code{TRUE}, includes standard errors in the table}

\item{se.inside}{if \code{TRUE} print standard errors in the same cell as estimates.}

\item{nobs}{if \code{TRUE}, print number of observations for each model.}

\item{digits}{Number of digits to round numbers to}

\item{rotate}{if \code{TRUE}, rows are models and columns are coefficients.}
}
\value{
an object of class \code{coeftab}
}
\description{
Returns a table of model coefficients in rows and models in columns. This can be used to compare estimates across models.
}
\examples{

data(WaffleDivorce)
d <- WaffleDivorce

# standardize variables
d$D <- standardize( d$Divorce )
d$M <- standardize( d$Marriage )
d$A <- standardize( d$MedianAgeMarriage )

m5.1 <- quap(
    alist(
        D ~ dnorm( mu , sigma ) ,
        mu <- a + bA * A ,
        a ~ dnorm( 0 , 0.2 ) ,
        bA ~ dnorm( 0 , 0.5 ) ,
        sigma ~ dexp( 1 )
    ) , data = d )

m5.2 <- quap(
alist(
    D ~ dnorm( mu , sigma ) ,
    mu <- a + bM * M ,
    a ~ dnorm( 0 , 0.2 ) ,
    bM ~ dnorm( 0 , 0.5 ) ,
    sigma ~ dexp( 1 )
) , data = d )

coeftab(m5.1,m5.2) 

}
