% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.r
\name{compare}
\alias{compare}
\alias{plot,compareIC,ANY-method}
\alias{compare.plot}
\alias{ICweights}
\alias{dSE,compareIC-method}
\alias{dSE}
\title{Compare fit models using WAIC or DIC}
\usage{
compare(
  ...,
  n = 1000,
  sort = "WAIC",
  func = WAIC,
  WAIC = TRUE,
  refresh = 0,
  warn = TRUE,
  result_order = c(1, 5, 3, 6, 2, 4)
)

\S4method{plot}{compareIC,ANY}(x, y, xlim, SE = TRUE, dSE = TRUE, weights = FALSE, ...)

ICweights(dev)

\S4method{dSE}{compareIC}(object)
}
\arguments{
\item{...}{A series of fit models, separated by commas.}

\item{n}{Number of samples from posterior to use in computing WAIC/DIC.}

\item{sort}{Sort table by ascending values in named column.}

\item{func}{Function to use in compting criteria for comparison.}

\item{WAIC}{Deprecated: If \code{TRUE}, uses \code{func} for comparison. If \code{FALSE}, uses DIC.}

\item{refresh}{Progress display update interval. 0 suppresses display.}

\item{x}{an object of type \code{compareIC} (output of \code{compare})}

\item{y}{not used, only for compatibility with \code{plot()}.}

\item{xlim}{the limits for the X axis}

\item{SE}{if \code{TRUE}, plots standard errors}

\item{dSE}{if \code{TRUE}, plots standard errors of differences.}

\item{weights}{if \code{TRUE}, convert estimated D_test values to weights.}

\item{dev}{Vector of values to use in computing model weights.}

\item{object}{an object of class \code{compareIC}}
}
\value{
An object of class \code{compareIC} with slots \code{.Data} (table of results) and \code{dSE} (matrix of standard errors of differences in IC between pairs of models)

for \code{dSE} the dSE matrix.
}
\description{
Returns a table of model comparison statistics, by default 
focused on WAIC.
}
\details{
This function computes WAIC and optionally DIC values for fit models and returns a table sorted by ascending values. Each row in this table is a model, and the various columns provide WAIC, effective numbers of parameters, model weights, and standard errors.

A \code{plot} method is supported, for graphic display of the information criteria.
}
\author{
Richard McElreath
}
