% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.r
\docType{class}
\name{compareIC-class}
\alias{compareIC-class}
\alias{show,compareIC-method}
\alias{compare}
\alias{ICweights}
\alias{plot,compareIC,ANY-method}
\title{Compare fit models using WAIC or DIC}
\usage{
\S4method{show}{compareIC}(object)

compare(..., n = 1000, sort = "WAIC", func = WAIC, WAIC = TRUE,
  refresh = 0)

\S4method{plot}{compareIC,ANY}(x, y, xlim, SE = TRUE, dSE = TRUE,
  weights = FALSE, ...)
}
\arguments{
\item{...}{A series of fit models, separated by commas}

\item{n}{Number of samples from posterior to use in computing WAIC/DIC}

\item{sort}{Sort table by ascending values in named column}

\item{func}{Function to use in computing criteria for comparison}

\item{WAIC}{Deprecated: If \code{TRUE}, uses \code{func} for comparison. If
\code{FALSE}, uses DIC.}

\item{refresh}{Progress display update interval. 0 suppresses display.}

\item{dev}{Vector of values to use in computing model weights}
}
\value{
An object of class \code{compareIC} with slots \code{output} (table
of results) and \code{dSE} (matrix of standard errors of differences in IC
between pairs of models).
}
\description{
Returns a table of model comparison statistics, by default focused on WAIC.
}
\details{
This function computes WAIC and optionally DIC values for fit models and
returns a table sorted by ascending values. Each row in this table is a
model, and the various columns provide WAIC, effective numbers of
parameters, model weights, and standard errors.

A \code{plot} method is supported, for graphic display of the information
criteria.
}
\author{
Richard McElreath
}
