\name{Crofoot}
\alias{Crofoot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Capuchin monkey contests}
\description{
  Data on features and outcomes of intergroup contests among territorial groups of capuchin monkey (Cebus capucinus). Each row contains a single contest between two groups.
}
\usage{
data(Crofoot)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
    \enumerate{
        \item focal : ID of focal group
        \item other : ID of other group
        \item dyad : ID of specific dyad pair of groups
        \item win : 1 if focal won contest, 0 if other won
        \item dist_focal : Distance in meters of focal group from the center of its home range
        \item dist_other : Distance in meters of other group from the center of its home range
        \item n_focal : Number of individuals in focal group
        \item n_other : Number of individuals in other group
        \item m_focal : Number of males in focal group
        \item m_other : Number of males in other group
        \item f_focal : Number of females in focal group
        \item f_other : Number of females in other group
    }
}
\value{
}
\references{M.C. Crofoot, I.C. Gilby, M.C. Wikelski, and R.W. Kays. 2008. PNAS 105:577--581.}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

