\name{Dissertations}
\alias{Dissertations}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Dissertation page length data}
\description{
  Page lengths of 3037 dissertations filed between 2006 and 2014 at the University of Minnesota.
}
\usage{
data(Dissertations)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
    \enumerate{
        \item pages: number of pages
        \item major: department/program of dissertation
        \item year: year of filing
    }
}
\value{
}
\references{Data originally parsed by...}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

