% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-class.r
\name{extract.samples}
\alias{extract.samples}
\title{Collect posterior samples from a fit model}
\usage{
extract.samples(object, n = 10000, clean.names = TRUE, ...)
}
\arguments{
\item{object}{Fit model to extract samples from}

\item{n}{Number of samples to simulate}

\item{clean.names}{When \code{TRUE}, removes parentheses from parameters
named \code{(Intercept)}.}

\item{...}{Other parameters to pass to descendent functions (when defined)}
}
\value{
A named \code{list} (for \code{map2stan}) or \code{data.frame}
containing samples for each parameter in the posterior distribution.
}
\description{
Extracts or draws posterior samples for fit models.
}
\details{
For \code{map2stan} models, this function returns cleaned samples already
contained in the object. These samples are cleaned of dimension attributes
and the \code{lp__} and \code{dev} traces that are used internally.

For \code{map} and other types, it uses the variance-covariance matrix and
coefficients to define a multivariate Gaussian posterior to draw \code{n}
samples from.
}
\seealso{
\code{\link{mvrnorm}}
}
\author{
Richard McElreath
}
