\name{Fish}
\alias{Fish}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Fishing data}
\description{
  Fishing data from visitors to a national park.
}
\usage{
data(Fish)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
    \enumerate{
        \item fish_caught : Number of fish caught during visit
        \item livebait : Whether or not group used livebait to fish
        \item camper : Whether or not group had a camper
        \item persons : Number of adults in group
        \item child : Number of children in group
        \item hours : Number of hours group spent in park
    }
}
\value{
}
\references{}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

