\name{foxes}
\alias{foxes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Urban foxes}
\description{
  Data on urban fox groups in England.
}
\usage{
data(foxes)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
    \enumerate{
        \item group : ID of group
        \item avgfood : Average available food in group's territory
        \item groupsize : Size of each group
        \item area : Area of group territory
        \item weight : Body weight of individual fox
    }
}
\value{
}
\references{Modified from an example in Grafen and Hails.}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

