\name{Hoogland}
\alias{Hoogland}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Prairie dog dispersal data}
\description{
  Dispersal and kin residence data for three species of prairie dog, from 1976 to 2004. Each row is an individual dispersal record, with associated descriptors.
}
\usage{
data(Hoogland)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
    \enumerate{
        \item Species : 1 = black-tailed, 2 = Gunnison's, 3 = Utah
        \item Year : Year of case record
        \item Male : 1 = male, 0 = female
        \item NoDisperse : 1 = did not disperse within 12 months of weaning, 0 = dispersed
        \item Mother : 1 = mother present for 12 months after weaning, 0 = absent
        \item Sisters : Minimal number of littermate sisters present at dispersal/non-dispersal
        \item Bros : Minimal number of littermate brothers present
        \item ClanSize : Minimal number of adults present in territory at time of dispersal/non-dispersal. Includes close kin, distant kin, immigrants, and focal individual
        \item AllKin : Minimal number of mother and littermates in territory at time of dispersal/non-dispersal
    }
}
\value{
}
\references{Hoogland 2013. Science 339:1205--1207.}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

