% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{HPDI}
\alias{HPDI}
\alias{PI}
\alias{PCI}
\title{Confidence/credible intervals from samples}
\usage{
HPDI(samples, prob = 0.89)
}
\arguments{
\item{samples}{Vector of parameter values}

\item{prob}{interval probability mass}
}
\description{
These functions compute highest posterior density (HPDI) and percentile (PI)
intervals, using samples from a posterior density or simulated outcomes.
}
\details{
Highest Posterior Density Intervals (HPDI) are calculated by
\code{\link{HPDinterval}} in the \code{coda} package.

Percentile intervals (PI) use \code{\link{quantile}()} and assign equal mass
to each tail.
}
\seealso{
\code{\link{HPDinterval}}
}
\author{
Richard McElreath
}
