% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcheck.r
\name{postcheck}
\alias{postcheck}
\title{Plot posterior predictive check}
\usage{
postcheck(fit, x, prob = 0.89, window = 20, n = 1000, col = rangi2, ...)
}
\arguments{
\item{fit}{A \code{map} or \code{map2stan} model fit}

\item{prob}{Width of prediction interval to display}

\item{window}{Number of cases to display in each plot (for paging)}

\item{n}{Number of samples from posterior to use}

\item{...}{Additional arguments to pass to \code{plot}}
}
\description{
Plots a series of graphical posterior predictive checks for a \code{map} or
\code{map2stan} model fit.
}
\details{
This function uses \code{\link{link}} and \code{\link{sim}} internally to
simulate posterior predictions for a fit model. These are then plotted over
the observed outcomes used in fitting.
}
\seealso{
\code{\link{link}}, \code{\link{sim}}
}
\author{
Richard McElreath
}
