% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rethinking.R
\docType{package}
\name{rethinking-package}
\alias{rethinking-package}
\alias{_PACKAGE}
\alias{rethinking}
\title{Statistical Rethinking package}
\description{
This package accompanies a book and course on Bayesian data
analysis, featured MAP estimation through \code{\link{quap}}
and Hamiltonian Monte Carlo through \code{\link{ulam}}.
}
\details{
\tabular{ll}{
Package: \tab rethinking\cr
Type: \tab Package\cr
License: \tab GPL-3 \cr
}
}
\references{
McElreath (2020). Statistical Rethinking: A Bayesian Course with Examples in R and Stan (2nd edition). CRC Press.
}
\seealso{
\code{\link{quap}}, \code{\link{ulam}}
}
\author{
Richard McElreath
}
\keyword{package}
\keyword{rethinking}
