\name{rethinking-package}
\alias{rethinking-package}
\alias{rethinking} 
\docType{package}
\title{Statistical Rethinking package}
\description{
  This package accompanies a book and course on Bayesian data analysis, featured MAP estimation through \code{\link{map}} and Hamiltonian Monte Carlo through \code{\link{map2stan}}. 
}

\details{
  \tabular{ll}{
    Package: \tab rethinking\cr
    Type: \tab Package\cr
    Version: \tab 1.56\cr
    Date: \tab 10 July 2015\cr
    License: \tab GPL-3 \cr
  }
  
}
\author{
  Richard McElreath
}
\references{
  McElreath (2016). Statistical Rethinking: A Bayesian Course with R Examples. CRC Press.
}

\examples{}  

\seealso{
  \code{\link{map}}, \code{\link{map2stan}} 
} 
\keyword{rethinking} 
\keyword{package} 