\name{rethinking-package}
\alias{rethinking-package}
\alias{rethinking} 
\docType{package}
\title{Statistical Rethinking package}
\description{
  This package accompanies a book and course on Bayesian data analysis, featured MAP estimation through \code{\link{quap}} and Hamiltonian Monte Carlo through \code{\link{ulam}}. 
}

\details{
  \tabular{ll}{
    Package: \tab rethinking\cr
    Type: \tab Package\cr
    License: \tab GPL-3 \cr
  }
  
}
\author{
  Richard McElreath
}
\references{
  McElreath (2020). Statistical Rethinking: A Bayesian Course with Examples in R and Stan (2nd edition). CRC Press.
}

\examples{}  

\seealso{
  \code{\link{quap}}, \code{\link{ulam}} 
} 
\keyword{rethinking} 
\keyword{package} 