\name{Rinder}
\alias{Rinder}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Livestock mortality data}
\description{
  Mortality counts for 11 herds of large and small livestock. Large livestock are Zebu cattle, and small livestock are a mix of (fat-tailed, Maasai) sheep and goats.
}
\usage{
data(Rinder)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
    \enumerate{
        \item herd : identifier for individual herd
        \item stock : categorical small or large stock
        \item n : number of animals at beginning of observation period
        \item mortality : number of animals that died in observation period
    }
}
\value{
}
\references{}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

