% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{simplehist}
\alias{simplehist}
\title{Simple histograms}
\usage{
simplehist(x, round = TRUE, ylab = "Frequency", ...)
}
\arguments{
\item{x}{Vector of values to construct histogram from}

\item{round}{When \code{TRUE}, rounds values in \code{x} before plotting}

\item{ylab}{Label on vertical axis}

\item{...}{Other parameters to pass to plot}
}
\description{
Simple integer-valued histograms, for displaying count distributions.
}
\details{
This function constructs clean histograms for count data. Non-integer data
can be rounded to nearest integer before plotting. Internally, this function
is little more than \code{plot(table(x))}.
}
\seealso{
\code{\link{hist}}
}
\author{
Richard McElreath
}
