\name{WaffleDivorce}
\alias{WaffleDivorce}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Waffle House and marriage statistics}
\description{
  Data for the individual States of the United States, describing number of Waffle House diners and various marriage and demographic facts.
}
\usage{
data(WaffleDivorce)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
    \enumerate{
        \item Location : State name
        \item Loc : State abbreviation
        \item Population : 2010 population in millions
        \item MedianAgeMarriage: 2005-2010 median age at marriage
        \item Marriage : 2009 marriage rate per 1000 adults
        \item Marriage.SE : Standard error of rate
        \item Divorce : 2009 divorce rate per 1000 adults
        \item Divorce.SE : Standard error of rate
        \item WaffleHouses : Number of diners
        \item South : 1 indicates Southern State
        \item Slaves1860 : Number of slaves in 1860 census
        \item Population1860 : Population from 1860 census
        \item PropSlaves1860 : Proportion of total population that were slaves in 1860
    }
}
\value{
}
\references{1860 census data from http://mapserver.lib.virginia.edu. Marriage and divorce rates from 2009 American Community Survey (ACS). Waffle House density data from wafflehouse.com (retrieved January 2012).}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

