\name{Wines2012}
\alias{Wines2012}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Wine tasting scores, Princeton NJ 2012}
\description{
  Numerical scores from 9 judges, both French and American, testing 20 different wines, both French and American. Judges were blind to the identity of each wine during the testing.
}
\usage{
data(Wines2012)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
    \enumerate{
        \item judge : Name of judge
        \item flight : white or red group of 10 wines
        \item wine : Wine ID
        \item score : Numerical score of wine
        \item wine.amer : 1 for American wines from New Jersey
        \item judge.amer : 1 for American judges
    }
}
\value{
}
\references{Raw data from http://www.liquidasset.com/report161.html}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

